backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis==5.2.1

# Auth
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.2.1

# Config
pydantic-settings==2.7.1
pydantic[email]==2.10.6
python-dotenv==1.0.1

# HTTP client
httpx==0.28.1

# AI / Vision
openai==1.58.1
anthropic==0.42.0

# Image processing
Pillow==11.1.0

# Nutrition API
# (using httpx for Nutritionix REST calls)

# Background tasks
apscheduler==3.10.4

# Encryption (for OAuth tokens)
cryptography==44.0.0

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx  # also used for test client

# Dev
black==24.10.0
ruff==0.8.6
